# linear search with numpy 10x faster