I tried doing the linear search using numpy and the results are 10x faster with slightly better scaling. This could be used until someone does a tensorflow implementation or a better search algorithm.

Feel free to leave it until you guys have time.

|number of memory observations|old speed (ms)|numpy (ms)|
|-|-|-|
|1e3|13.9 |0.9|
|1e4|132.0 |4.7| 
|1e5|1690.0 |55.6| 

Related discussion: https://github.com/reinforceio/tensorforce/pull/164#issuecomment-336389944