@AlexKuhnle 2017-10-13
That looks pretty good, thanks!