@wassname 2017-10-14
Thanks, I'm closing in favour of a better version #168